+++ OUDS/Core/ThemesContract/Sources/TokenatorConstants.swift
@@ -41,10 +41,10 @@ public enum OUDSVersions {
 
     // MARK: - Components versions - Actions
 
-    /// Version of the Figma specifications for the component button (3.4.0)
-    public static let componentButtonVersion = "3.4.0"
-    /// Version of the Figma specifications for the component interactive icon (1.1.0)
-    public static let componentInteractiveIcon = "1.1.0"
+    /// Version of the Figma specifications for the component button (3.3.0)
+    public static let componentButtonVersion = "3.3.0"
+    /// Version of the Figma specifications for the component interactive icon (1.0.0)
+    public static let componentInteractiveIcon = "1.0.0"
 
     // MARK: - Components versions - AI assistant
 
@@ -82,14 +82,14 @@ public enum OUDSVersions {
     public static let componentPinCodeInputVersion = "1.3.0"
     /// Version of the Figma specifications for the component quantity input (1.3.1)
     public static let componentQuantityInputVersion = "1.3.1"
-    /// Version of the Figma specifications for the component quantity input (1.2.0)
-    public static let componentStaticListItemVersion = "1.2.0"
-    /// Version of the Figma specifications for the component quantity input (1.2.0)
-    public static let componentStaticCardItemVersion = "1.2.0"
-    /// Version of the Figma specifications for the component quantity input (1.2.0)
-    public static let componentNavigationListItemVersion = "1.2.0"
-    /// Version of the Figma specifications for the component quantity input (1.2.0)
-    public static let componentNavigationCardItemVersion = "1.2.0"
+    /// Version of the Figma specifications for the component quantity input (1.0.0)
+    public static let componentStaticListItemVersion = "1.0.0"
+    /// Version of the Figma specifications for the component quantity input (1.0.0)
+    public static let componentStaticCardItemVersion = "1.0.0"
+    /// Version of the Figma specifications for the component quantity input (1.0.0)
+    public static let componentNavigationListItemVersion = "1.0.0"
+    /// Version of the Figma specifications for the component quantity input (1.0.0)
+    public static let componentNavigationCardItemVersion = "1.0.0"
     /// Version of the Figma specifications for the component radio button (1.4.0)
     public static let componentRadioButtonVersion = "1.4.0"
     /// Version of the Figma specifications for the component select input (1.3.1)
@@ -103,8 +103,8 @@ public enum OUDSVersions {
 
     // MARK: - Components versions - Dialog
 
-    /// Version of the Figma specifications for the component alert (alert message) (1.2.0)
-    public static let componentAlertMessageVersion = "1.2.0"
+    /// Version of the Figma specifications for the component alert (alert message) (1.1.1)
+    public static let componentAlertMessageVersion = "1.1.1"
     /// Version of the Figma specifications for the component alert (inline alert) (1.0.0)
     public static let componentInlineAlertVersion = "1.0.0"
     /// Version of the Figma specifications for the component alert (alert bannet) (1.0.0)
@@ -114,8 +114,8 @@ public enum OUDSVersions {
 
     // MARK: - Components versions - Foundations
 
-    /// Version of the Figma specifications for the component typgography) (1.1.0)
-    public static let componentTypographyVersion = "1.1.0"
+    /// Version of the Figma specifications for the component typgography) (1.0.0)
+    public static let componentTypographyVersion = "1.0.0"
 
     // MARK: - Components versions - Indicator